rootfs: add hook script to be called by dhclient-script
The dns servers will be needed later on to get the ip address of enfield which
provides the early-boot config values.

Unfortunately, since /etc is currently read-only, makeresolveconf in
dhclient-script will fail to write the dns servers to disk. So we hook into
dhclient-script when exiting to save the provided dns servers.

The dhcp server is also saved just as a nice-to-have.