Added the ability to build a go package using source from a location outside the normal buildroot tree. This is driven by the new script 'buildgopackage' which sets some environment variables and then uses the 'buildmistify' script to build. At the moment, to avoid some dependency problems, the 'buildmistify' script is invoked twice. Once to remove the previous sources from the buildroot build tree and the second time to build the package.
To tie this together a new package named 'gopackage' has been created.
It uses a package name environment variable as the version which enables
building different packages using this scheme. NOTE: it is best to use
'buildgopackage' because of the environment variable dependencies.

Lastly, a sample package has been added (this location may be temporary
because it slightly pollutes the mistify-os project -- however, it
does make some sense because of the purpose of mistify-os). This package
was derived from a test subagent under the mistify-agent. It builds and
installs but doesn't have the appropriate scripts to configure mistify-agent
to use it.

NOTE: This is a Work In Progress but is being pushed to the repo for
comments.